My project on Data analysis and Categorization
The project may have limited content since it focuses on animal data. The automatic analysis and categorization of the data, combined with livestock management tools, can help farmers make decisions. This system, which will allow to collect, document, and use data to stores animal records according to the organization's needs. It enables users to use already existing sets of data, manipulate the database accordingly, and access information as they see fit with a framework.